+++ ott/math/__init__.py
@@ -3,7 +3,6 @@ from . import (
     fixed_point_loop,
     implicit_differentiation,
     matrix_square_root,
-    potentials,
     unbalanced_functions,
     utils,
 )
