+++ ott/problems/linear/potentials.py
