+++ ott/typing.py
@@ -1,4 +1,4 @@
-from jax import numpy as jnp
+import jax.numpy as jnp
 from typing_extensions import Protocol
 
 # TODO(michalk8): introduce additional types here
