+++ ott/utils.py
