--- ott/utils/__init__.py
@@ -1,1 +1,0 @@
-from . import dataclasses, segment, was_solver
